+++ src/utils/Token.sol
@@ -5,13 +5,9 @@ import {ERC20} from "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
 
 contract Token is ERC20, Ownable {
-    uint public initialSupply = 10_000_000 ether;
+    uint256 public initialSupply = 10_000_000 ether;
 
-    constructor(
-        string memory _name,
-        string memory _symbol,
-        address _owner
-    ) ERC20(_name, _symbol) Ownable(_owner) {
+    constructor(string memory _name, string memory _symbol, address _owner) ERC20(_name, _symbol) Ownable(_owner) {
         _mint(_owner, initialSupply);
     }
 }
